TDCHit: fix invalid extended trigger time tag storage and interpretation
Extended trigger time tag is not more significant bits of a time
measurement. Instead, it is the time of the beginning of the trigger
window sampled from a clock of lower frequency (25 ns vs 25 pcs). Also
its least significant bits are stored in the trailer GEO address.

Delete
  TDCHit::GetValue
  TDCHit::SetValue
Rename
  TDCHit::GetMeasurement -> TDCHit::GetValue
  TDCHit::SetMeasurement -> TDCHit::SetValue

TDCHit byte structure is almost completely rearranged and is not
backwards compatible